+++ src/Resources/app/administration/src/module/muwa-backup-repository/page/muwa-backup-repository-list/index.js
@@ -11,7 +11,8 @@ Component.register('muwa-backup-repository-list', {
     template,
 
     inject: [
-        'repositoryFactory'
+        'repositoryFactory',
+        'acl'
     ],
 
     mixins: [